+++ authentication/models.py
@@ -41,19 +41,3 @@ class User(AbstractUser):
         return self.email
 
 
-class Expense(models.Model):
-    CATEGORY_CHOICES = [
-        ('Groceries', 'Groceries'),
-        ('Leisure', 'Leisure'),
-        ('Utilities', 'Utilities'),
-        ('Transport', 'Transport'),
-        ('Other', 'Other'),
-    ]
-
-    amount = models.DecimalField(max_digits=10, decimal_places=2)
-    category = models.CharField(max_length=20, choices=CATEGORY_CHOICES)
-    description = models.TextField(blank=True, null=True)
-    date = models.DateField(auto_now_add=True)
-
-    def __str__(self):
-        return f" {self.category} - ${self.amount}"
