+++ config/urls.py
@@ -28,6 +28,7 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     path('api/', include('authentication.urls')),
     path("api/auth/", include("authentication.urls")),
+    path('api/expenses/', include('expense.urls')),
     # OpenAPI Schema
     path("api/schema/", SpectacularAPIView.as_view(), name="schema"),
     # Swagger UI
