+++ expense/admin.py
@@ -1,0 +1,9 @@
+from django.contrib import admin
+from expense.models import Expense
+
+
+# Register your models here.
+admin.site.register(Expense)
+
+
+
