+++ expense/serializers.py
@@ -1,8 +1,8 @@
 from rest_framework import serializers
-from .models import Expense
+from expense.models import Expense
 
 class ExpenseSerializer(serializers.ModelSerializer):
     class Meta:
         model = Expense
-        fields = ['id',  'amount', 'category', 'description', 'date']
+        fields =  '__all__'
